Move monitoring files to their own submodule, update CMakeLists
The decision on whether to include monitoring is done based on the availability of ROOT.
EUDAQ bases its monitor off ROOT's GUI system, so if ROOT is available, we assume that EUDAQ was built with its ROOT-based monitor enabled and we include the monitor submodule. If ROOT is unavailable, the monitor submodule is ignored.