skip events with ticks since spill above 2^18
this is probably undesirable and may need more configuration in the
future but it is a first-pass at a "quality" cut for each event. we know
a priori that the spills are very short so the event ticks within a
spill should be grouped together

you can see this with a plot of ticks vs event number for a single
un-aligned polarfire. you see a sawtooth pattern and occasionally a
outlier crops up ruining the pretty picture